Moved files from core into backend folder and deleted core. Updated imports
