+++ elevate/src/backend/CoreStateManager.ts
