+++ elevate/src/backend/FileSnapshot.ts
